docs-daemon-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctl-sync-daemon"
version = "3.0.0"
description = "CTL v3.0 Tensor Language Sync Daemon (Python)"
authors = [{name = "MAGRAY_CLI", email = "magray@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "watchdog>=3.0.0",
    "jsonschema>=4.19.2",
    "pydantic>=2.5.0",
    "click>=8.1.7",
    "PyYAML>=6.0.1",
    "colorama>=0.4.6",
]

[project.scripts]
ctl-sync = "ctl_sync.main:main"

[project.urls]
Homepage = "https://github.com/MAGRAY_CLI/MAGRAY_Cli"
Repository = "https://github.com/MAGRAY_CLI/MAGRAY_Cli"
Issues = "https://github.com/MAGRAY_CLI/MAGRAY_Cli/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ctl_sync*"]
exclude = ["tests*"]